ex 4.15 - 4.23: token authentication, fixin tests

+++ part4/blog/api/src/controllers/blog.controller.js
@@ -1,28 +1,33 @@
 const Blog = require('../models/Blog');
 
 const getAllBlogs = async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { username: 1 });
   res.json(blogs);
 };
 
-const addBlog = async (req, res) => {
-  const { body } = req;
+const addBlog = async (req, res, next) => {
+  const { author, likes, title, url } = req.body;
+  const { userFromDB } = req;
 
-  if (!body.title || !body.url)
+  if (!title || !url) {
     return res.status(400).json({ error: 'Title and url must not be empty.' });
+  }
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
+    title,
+    author,
+    url,
+    likes: likes || 0,
+    user: userFromDB._id,
   });
 
   try {
     const savedBlog = await blog.save();
+    userFromDB.blogs = [...userFromDB.blogs, savedBlog._id];
+    await userFromDB.save();
     res.status(201).json(savedBlog);
   } catch (err) {
-    res.status(400).json({ error: err });
+    next(err);
   }
 };
 
@@ -46,13 +51,27 @@ const updateBlog = async (req, res) => {
   }
 };
 
-const deleteBlog = async (req, res) => {
+const deleteBlog = async (req, res, next) => {
   const { id } = req.params;
+  const { userFromDB } = req;
   try {
+    const blogFromDB = await Blog.findById(id);
+
+    if (blogFromDB.user.toString() !== userFromDB._id.toString()) {
+      return res
+        .status(401)
+        .json({ error: 'You cannot delete this blog. Wrong credentials.' });
+    }
+
+    userFromDB.blogs = userFromDB.blogs.filter(
+      blog => blog._id.toString() !== id.toString()
+    );
+    await userFromDB.save();
+
     await Blog.findByIdAndRemove(id);
     res.status(204).end();
   } catch (err) {
-    res.status(400).json({ error: err });
+    next(err);
   }
 };
